moved implementation of getMode to sci_size.cpp
