+++ scilab/modules/elementary_functions/sci_gateway/cpp/sci_size.cpp
@@ -2,8 +2,8 @@
  * Scilab ( http://www.scilab.org/ ) - This file is part of Scilab
  * Copyright (C) 2012 - DIGITEO - Antoine ELIAS
  * Copyright (C) 2015 - Scilab Enterprises - Anais AUBERT
- *
  * Copyright (C) 2012 - 2016 - Scilab Enterprises
+ * Copyright (C) 2017 - 2018 Dirk Reusch, Kybernetik Dr. Reusch
  *
  * This file is hereby licensed under the terms of the GNU GPL v2.0,
  * pursuant to article 5.3.4 of the CeCILL v.2.1.
@@ -13,15 +13,14 @@
  * along with this program.
  *
  */
-/*--------------------------------------------------------------------------*/
 
 #include <algorithm>
 
 #include "elem_func_gw.hxx"
 #include "types.hxx"
+#include "double.hxx"
 #include "string.hxx"
 #include "container.hxx"
-#include "getmode.hxx"
 #include "overload.hxx"
 #include "context.hxx"
 
@@ -32,7 +31,8 @@ extern "C"
 #include "os_string.h"
 }
 
-/*--------------------------------------------------------------------------*/
+static int getMode(types::typed_list &in, int _iProcess, int _iRef);
+
 types::Function::ReturnValue sci_size(types::typed_list &in, int _iRetCount, types::typed_list &out)
 {
     if (in.size() < 1)
@@ -175,4 +175,81 @@ types::Function::ReturnValue sci_size(types::typed_list &in, int _iRetCount, typ
     }
     return types::Function::OK;
 }
-/*--------------------------------------------------------------------------*/
+
+int getMode(types::typed_list &in, int _iProcess, int _iRef)
+{
+    int iMode = 0;
+    if (in[_iProcess]->isString())
+    {
+        types::String* pS = in[_iProcess]->getAs<types::String>();
+        if (pS->getSize() != 1)
+        {
+            Scierror(999, _("%s: Wrong size for argument %d: (%d,%d) expected.\n"), "size", _iProcess + 1, 1, 1);
+        }
+
+        switch (pS->getFirst()[0])
+        {
+            case 'r' :
+                iMode = 1;
+                break;
+            case 'c' :
+                iMode = 2;
+                break;
+            case '*' :
+                iMode = 0;
+                break;
+            case 'm' :
+                iMode = -1;
+                break;
+            default :
+                Scierror(999, _("%s: Wrong value for input argument #%d: '%s', '%s', '%s' or '%s' expected.\n"), "size", _iProcess + 1, "m" , "*" , "r", "c");
+                iMode = -2;
+                break;
+        }
+    }
+    else if (in[1]->isDouble() && in[1]->getAs<types::Double>()->isComplex() == false)
+    {
+        types::Double* pD = in[_iProcess]->getAs<types::Double>();
+        if (pD->getSize() != 1)
+        {
+            Scierror(999, _("%s: Wrong size for argument %d: (%d,%d) expected.\n"), "size", _iProcess + 1, 1, 1);
+            iMode = -2;
+        }
+        else
+        {
+            iMode = static_cast<int>(pD->getReal()[0]);
+            if (pD->getReal()[0] != static_cast<double>(iMode))
+            {
+                Scierror(999, _("%s: Wrong value for input argument #%d: An integer value expected.\n"), "size", 2);
+                iMode = -2;
+            }
+
+            if (iMode <= 0)
+            {
+                Scierror(999, _("%s: Wrong value for input argument #%d: Scalar positive integer expected.\n"), "size", 2);
+                iMode = -2;
+            }
+        }
+    }
+    else
+    {
+        Scierror(999, _("%s: Wrong type for input argument #%d: string or scalar expected.\n"), "size", 2);
+        iMode = -2;
+    }
+
+    //special case for -1
+    if (iMode == -1)
+    {
+        iMode = 0;
+        if (in[_iRef]->getAs<types::GenericType>()->getRows() > 1)
+        {
+            iMode = 1;
+        }
+        else if (in[_iRef]->getAs<types::GenericType>()->getCols() > 1)
+        {
+            iMode = 2;
+        }
+    }
+
+    return iMode;
+}
