+++ scilab/modules/javasci/Makefile.in
@@ -1018,10 +1018,10 @@ maintainer-clean-generic:
 	@echo "This command is intended for maintainers to use"
 	@echo "it deletes files that may require special tools to rebuild."
 	-test -z "$(BUILT_SOURCES)" || rm -f $(BUILT_SOURCES)
+@JAVASCI_FALSE@install-data-local:
 @JAVASCI_FALSE@install-html-local:
 @JAVASCI_FALSE@clean-local:
 @JAVASCI_FALSE@distclean-local:
-@JAVASCI_FALSE@install-data-local:
 clean: clean-am
 
 clean-am: clean-generic clean-libtool clean-local \
